Move the safe handles to a separate namespace.
