+++ sources/TileDB.CSharp/Config.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Text;
+using TileDB.CSharp.Marshalling.SafeHandles;
 using TileDB.Interop;
 
 namespace TileDB.CSharp
