+++ sources/TileDB.CSharp/Domain.cs
@@ -1,4 +1,5 @@
 using System;
+using TileDB.CSharp.Marshalling.SafeHandles;
 using TileDB.Interop;
 
 namespace TileDB.CSharp
